+++ examples/ThirdPartyLibs/BussIK/LinearR2.h
@@ -353,8 +353,8 @@ inline VectorR2& VectorR2::Load(const double* v)
 
 inline VectorR2& VectorR2::Load(const float* v)
 {
-	x = *v;
-	y = *(v + 1);
+	x = (double)*v;
+	y = (double)*(v + 1);
 	return *this;
 }
